correcciones no normineteado

+++ printf.h
@@ -17,14 +17,14 @@
 # include <stdlib.h>
 # include <unistd.h>
 
-int	ft_printf(const char *format, ...);
-int	ft_what_to_print(va_list args, char c);
-int	ft_print_c(int c);
-int	ft_print_s(char *str);
-int	ft_print_p(void *ptr);
-int	ft_print_d(int n);
-int	ft_print_u(unsigned int n);
-int	ft_print_x(unsigned int n, char *base);
+int		ft_printf(const char *format, ...);
+int		ft_what_to_print(va_list args, char c);
+int		ft_print_c(int c);
+int		ft_print_s(char *str);
+int		ft_print_p(void *ptr);
+int		ft_print_d(int n);
+int		ft_print_u(unsigned int n);
+int		ft_print_x(unsigned long n, char *base);
 void	ft_putchar(int c);
 
 #endif
